Rename files (phase II).
